moved global/component bootstrap code to a module, removed react-router/redux from App
